@@ -1,6 +1,6 @@
 [metadata]
 name = dials_data
-version = 2.2.0
+version = 2.4.0
 url = https://github.com/dials/data
 project_urls =
     Bug Tracker = https://github.com/dials/data/issues
@@ -19,7 +19,6 @@ classifiers =
     License :: OSI Approved :: BSD License
     Natural Language :: English
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -32,9 +31,10 @@ install_requires =
     importlib_resources>=1.1
     pytest
     pyyaml
+    requests
 # importlib; python_version == "2.6"
 packages = find:
-python_requires = >=3.6
+python_requires = >=3.7
 zip_safe = False
 
 [options.entry_points]
@@ -46,6 +46,9 @@ libtbx.precommit =
     dials_data = dials_data
 pytest11 =
     dials_data = dials_data.pytest11
+
+[options.package_data]
+dials_data = py.typed
 
 [flake8]
 # Black disagrees with flake8 on a few points. Ignore those.
