@@ -3,9 +3,11 @@ DIALS Regression Data Manager
 https://github.com/dials/data
 """
 
+from __future__ import annotations
+
 __all__: list = []
 __author__ = """Markus Gerstel"""
 __email__ = "dials-support@lists.sourceforge.net"
-__version__ = "2.2.0"
+__version__ = "2.4.0"
 __commit__ = ""
 __version_tuple__ = tuple(int(x) for x in __version__.split("."))
