@@ -1,12 +1,22 @@
+from __future__ import annotations
+
+import concurrent.futures
 import contextlib
 import errno
+import functools
+import hashlib
 import os
 import tarfile
 import warnings
 import zipfile
 from pathlib import Path
+from typing import Any, Optional, Union
 from urllib.parse import urlparse
-from urllib.request import urlopen
+
+import py.path
+import requests
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
 
 import dials_data.datasets
 
@@ -18,7 +28,6 @@ if os.name == "posix":
 
     def _platform_unlock(file_handle):
         fcntl.lockf(file_handle, fcntl.LOCK_UN)
-
 
 elif os.name == "nt":
     import msvcrt
@@ -38,7 +47,6 @@ elif os.name == "nt":
     def _platform_unlock(file_handle):
         file_handle.seek(0)
         msvcrt.locking(file_handle.fileno(), msvcrt.LK_UNLCK, 1)
-
 
 else:
 
@@ -68,69 +76,46 @@ def _file_lock(file_handle):
             _platform_unlock(file_handle)
 
 
-@contextlib.contextmanager
-def download_lock(target_dir):
-    """
-    Obtains a (cooperative) lock on a lockfile in a target directory, so only a
-    single (cooperative) process can enter this context manager at any one time.
-    If the lock is held this will block until the existing lock is released.
-    """
-    with target_dir.join(".lock").open(mode="w", ensure=True) as fh:
-        with _file_lock(fh):
-            yield
-
-
-def _download_to_file(url, pyfile):
+def _download_to_file(session: requests.Session, url: str, pyfile: Path):
     """
     Downloads a single URL to a file.
     """
-    with contextlib.closing(urlopen(url)) as socket:
-        file_size = socket.info().get("Content-Length")
-        if file_size:
-            file_size = int(file_size)
-        # There is no guarantee that the content-length header is set
-        received = 0
-        block_size = 8192
-        # Allow for writing the file immediately so we can empty the buffer
-        with pyfile.open(mode="wb", ensure=True) as f:
-            while True:
-                block = socket.read(block_size)
-                received += len(block)
-                f.write(block)
-                if not block:
-                    break
-
-    if file_size and file_size != received:
-        raise OSError(
-            "Error downloading {url}: received {received} bytes instead of expected {file_size} bytes".format(
-                file_size=file_size, received=received, url=url
-            )
-        )
+    with session.get(url, stream=True) as r:
+        r.raise_for_status()
+        pyfile.parent.mkdir(parents=True, exist_ok=True)
+        with pyfile.open(mode="wb") as f:
+            for chunk in r.iter_content(chunk_size=40960):
+                f.write(chunk)
 
 
-def file_hash(path):
+def file_hash(file_to_hash: Path) -> str:
     """Returns the SHA256 digest of a file."""
-    return path.computehash(hashtype="sha256")
+    sha256_hash = hashlib.sha256()
+    with file_to_hash.open("rb") as f:
+        for block in iter(lambda: f.read(4096), b""):
+            sha256_hash.update(block)
+    return sha256_hash.hexdigest()
 
 
 def fetch_dataset(
     dataset,
-    ignore_hashinfo=False,
-    verify=False,
-    read_only=False,
-    verbose=False,
-    pre_scan=True,
-    download_lockdir=None,
-):
+    ignore_hashinfo: bool = False,
+    verify: bool = False,
+    read_only: bool = False,
+    verbose: bool = False,
+    pre_scan: bool = True,
+) -> Union[bool, dict[str, Any]]:
     """Check for the presence or integrity of the local copy of the specified
     test dataset. If the dataset is not available or out of date then attempt
     to download/update it transparently.
 
-    :param verbose:          Show everything as it happens.
     :param pre_scan:         If all files are present and all file sizes match
                              then skip file integrity check and exit quicker.
     :param read_only:        Only use existing data, never download anything.
                              Implies pre_scan=True.
+    :param verbose:          Show everything as it happens.
+    :param verify:           Check all files against integrity information and
+                             fail on any mismatch.
     :returns:                False if the dataset can not be downloaded/updated
                              for any reason.
                              True if the dataset is present and passes a
@@ -142,8 +127,8 @@ def fetch_dataset(
         return False
     definition = dials_data.datasets.definition[dataset]
 
-    target_dir = dials_data.datasets.repository_location().join(dataset)
-    if read_only and not target_dir.check(dir=1):
+    target_dir: Path = dials_data.datasets.repository_location() / dataset
+    if read_only and not target_dir.is_dir():
         return False
 
     integrity_info = definition.get("hashinfo")
@@ -152,10 +137,10 @@ def fetch_dataset(
 
     if "verify" not in integrity_info:
         integrity_info["verify"] = [{} for _ in definition["data"]]
-    filelist = [
+    filelist: list[dict[str, Any]] = [
         {
             "url": source["url"],
-            "file": target_dir.join(os.path.basename(urlparse(source["url"]).path)),
+            "file": target_dir / os.path.basename(urlparse(source["url"]).path),
             "files": source.get("files"),
             "verify": hashinfo,
         }
@@ -164,90 +149,118 @@ def fetch_dataset(
 
     if pre_scan or read_only:
         if all(
-            item["file"].check()
+            item["file"].is_file()
             and item["verify"].get("size")
-            and item["verify"]["size"] == item["file"].size()
+            and item["verify"]["size"] == item["file"].stat().st_size
             for item in filelist
         ):
             return True
         if read_only:
             return False
 
-    if download_lockdir:
-        # Acquire lock if required as files may be downloaded/written.
-        with download_lock(download_lockdir):
-            _fetch_filelist(filelist, file_hash)
-    else:
-        _fetch_filelist(filelist, file_hash)
+    # Obtain a (cooperative) lock on a dataset-specific lockfile, so only one
+    # (cooperative) process can enter this context at any one time. The lock
+    # file sits in the directory above the dataset directory, as to not
+    # interfere with dataset files.
+    target_dir.mkdir(parents=True, exist_ok=True)
+    with target_dir.with_name(f".lock.{dataset}").open(mode="w") as fh:
+        with _file_lock(fh):
+            verification_records = _fetch_filelist(filelist)
 
+    # If any errors occured during download then don't trust the dataset.
+    if verify and not all(verification_records):
+        return False
+
+    integrity_info["verify"] = verification_records
     return integrity_info
 
 
-def _fetch_filelist(filelist, file_hash):
-    for source in filelist:  # parallelize this
-        if source.get("type", "file") == "file":
+def _fetch_filelist(filelist: list[dict[str, Any]]) -> list[dict[str, Any] | None]:
+    with requests.Session() as rs:
+        retry_adapter = HTTPAdapter(
+            max_retries=Retry(
+                total=5,
+                backoff_factor=1,
+                raise_on_status=True,
+                status_forcelist={413, 429, 500, 502, 503, 504},
+            )
+        )
+        rs.mount("http://", retry_adapter)
+        rs.mount("https://", retry_adapter)
+
+        pool = concurrent.futures.ThreadPoolExecutor(max_workers=5)
+        results = pool.map(functools.partial(_fetch_file, rs), filelist)
+        return list(results)
+
+
+def _fetch_file(
+    session: requests.Session, source: dict[str, Any]
+) -> dict[str, Any] | None:
+    valid = False
+    if source["file"].is_file():
+        # verify
+        valid = True
+        if source["verify"]:
+            if source["verify"]["size"] != source["file"].stat().st_size:
+                valid = False
+            elif source["verify"]["hash"] != file_hash(source["file"]):
+                valid = False
+
+    downloaded = False
+    if not valid:
+        print(f"Downloading {source['url']}")
+        _download_to_file(session, source["url"], source["file"])
+        downloaded = True
+
+    # verify
+    validation_record = {
+        "size": source["file"].stat().st_size,
+        "hash": file_hash(source["file"]),
+    }
+    valid = True
+    if source["verify"]:
+        if source["verify"]["size"] != validation_record["size"]:
+            print(
+                f"File size mismatch on {source['file']}: "
+                f"{validation_record['size']}, expected {source['verify']['size']}"
+            )
             valid = False
-            if source["file"].check(file=1):
-                # verify
-                valid = True
-                if source["verify"]:
-                    if source["verify"]["size"] != source["file"].size():
-                        valid = False
-                        print("size")
-                    elif source["verify"]["hash"] != file_hash(source["file"]):
-                        valid = False
+        elif source["verify"]["hash"] != validation_record["hash"]:
+            print(f"File hash mismatch on {source['file']}")
+            valid = False
+
+    # If the file is a tar archive, then decompress
+    if source["files"]:
+        target_dir = source["file"].parent
+        if downloaded or not all((target_dir / f).is_file() for f in source["files"]):
+            # If the file has been (re)downloaded, or we don't have all the requested
+            # files from the archive, then we need to decompress the archive
+            print(f"Decompressing {source['file']}")
+            if source["file"].suffix == ".zip":
+                with zipfile.ZipFile(source["file"]) as zf:
+                    try:
+                        for f in source["files"]:
+                            zf.extract(f, path=source["file"].parent)
+                    except KeyError:
                         print(
-                            "hash", source["verify"]["hash"], file_hash(source["file"])
+                            f"Expected file {f} not present "
+                            f"in zip archive {source['file']}"
                         )
-
-            downloaded = False
-            if not valid:
-                print("Downloading {}".format(source["url"]))
-                _download_to_file(source["url"], source["file"])
-                downloaded = True
-
-            # verify
-            valid = True
-            fileinfo = {
-                "size": source["file"].size(),
-                "hash": file_hash(source["file"]),
-            }
-            if source["verify"]:
-                if source["verify"]["size"] != fileinfo["size"]:
-                    valid = False
-                elif source["verify"]["hash"] != fileinfo["hash"]:
-                    valid = False
             else:
-                source["verify"]["size"] = fileinfo["size"]
-                source["verify"]["hash"] = fileinfo["hash"]
-
-        # If the file is a tar archive, then decompress
-        if source["files"]:
-            target_dir = source["file"].dirpath()
-            if downloaded or not all(
-                (target_dir / f).check(file=1) for f in source["files"]
-            ):
-                # If the file has been (re)downloaded, or we don't have all the requested
-                # files from the archive, then we need to decompress the archive
-                print("Decompressing {file}".format(file=source["file"]))
-                if source["file"].ext == ".zip":
-                    with zipfile.ZipFile(source["file"].strpath) as zf:
+                with tarfile.open(source["file"]) as tar:
+                    for f in source["files"]:
                         try:
-                            for f in source["files"]:
-                                zf.extract(f, path=source["file"].dirname)
+                            tar.extract(f, path=source["file"].parent, set_attrs=False)
                         except KeyError:
                             print(
-                                f"Expected file {f} not present in zip archive {source['file']}"
+                                f"Expected file {f} not present "
+                                f"in tar archive {source['file']}"
                             )
-                else:
-                    with tarfile.open(source["file"].strpath) as tar:
-                        for f in source["files"]:
-                            try:
-                                tar.extract(f, path=source["file"].dirname)
-                            except KeyError:
-                                print(
-                                    f"Expected file {f} not present in tar archive {source['file']}"
-                                )
+
+    if valid:
+        return validation_record
+    else:
+        return None
 
 
 class DataFetcher:
@@ -257,7 +270,7 @@ class DataFetcher:
         df = DataFetcher()
     Then
         df('insulin')
-    returns a py.path object to the insulin data. If that data is not already
+    returns a Path object to the insulin data. If that data is not already
     on disk it is downloaded automatically.
 
     To disable all downloads:
@@ -267,14 +280,14 @@ class DataFetcher:
     """
 
     def __init__(self, read_only=False):
-        self._cache = {}
-        self._target_dir = dials_data.datasets.repository_location()
-        self._read_only = read_only and os.access(self._target_dir.strpath, os.W_OK)
+        self._cache: dict[str, Optional[Path]] = {}
+        self._target_dir: Path = dials_data.datasets.repository_location()
+        self._read_only: bool = read_only and os.access(self._target_dir, os.W_OK)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return "<{}DataFetcher: {}>".format(
             "R/O " if self._read_only else "",
-            self._target_dir.strpath,
+            self._target_dir,
         )
 
     def result_filter(self, result, **kwargs):
@@ -286,7 +299,7 @@ class DataFetcher:
         """
         return result
 
-    def __call__(self, test_data, pathlib=None, **kwargs):
+    def __call__(self, test_data: str, pathlib=None, **kwargs):
         """
         Return the location of a dataset, transparently downloading it if
         necessary and possible.
@@ -302,7 +315,7 @@ class DataFetcher:
                  if the dataset is not available.
         """
         if test_data not in self._cache:
-            self._cache[test_data] = self._attempt_fetch(test_data, **kwargs)
+            self._cache[test_data] = self._attempt_fetch(test_data)
         if pathlib is None:
             warnings.warn(
                 "The DataFetcher currently returns py.path.local() objects. "
@@ -312,15 +325,13 @@ class DataFetcher:
                 DeprecationWarning,
                 stacklevel=2,
             )
-        if pathlib and self._cache[test_data]["result"]:
-            result = {
-                **self._cache[test_data],
-                "result": Path(self._cache[test_data]["result"]),
-            }
-            return self.result_filter(**result)
-        return self.result_filter(**self._cache[test_data])
+        if not self._cache[test_data]:
+            return self.result_filter(result=False)
+        elif not pathlib:
+            return self.result_filter(result=py.path.local(self._cache[test_data]))
+        return self.result_filter(result=self._cache[test_data])
 
-    def _attempt_fetch(self, test_data):
+    def _attempt_fetch(self, test_data: str) -> Optional[Path]:
         if self._read_only:
             data_available = fetch_dataset(test_data, pre_scan=True, read_only=True)
         else:
@@ -328,9 +339,8 @@ class DataFetcher:
                 test_data,
                 pre_scan=True,
                 read_only=False,
-                download_lockdir=self._target_dir,
             )
         if data_available:
-            return {"result": self._target_dir.join(test_data)}
+            return self._target_dir / test_data
         else:
-            return {"result": False}
+            return None
